+++ lib/native_geofence.dart
@@ -1,4 +1,4 @@
-export 'package:native_geofence/src/generated/platform_bindings.g.dart'
+export 'package:native_geofence/src/generated/native_geofence_api.g.dart'
     show
         Geofence,
         Location,
@@ -9,6 +9,8 @@ export 'package:native_geofence/src/generated/platform_bindings.g.dart'
         GeofenceStatus,
         NativeGeofenceErrorCode,
         GeofenceCallbackParams;
+export 'package:native_geofence/src/generated/native_geofence_api.g.dart'
+    show ForegroundServiceConfiguration;
 export 'package:native_geofence/src/model/native_geofence_exception.dart';
 export 'package:native_geofence/src/native_geofence_background_manager.dart'
     show NativeGeofenceBackgroundManager;