+++ lib/src/model/native_geofence_exception.dart
@@ -1,4 +1,6 @@
-import 'package:native_geofence/src/generated/platform_bindings.g.dart';
+
+
+import '../generated/native_geofence_api.g.dart';
 
 class NativeGeofenceException implements Exception {
   final NativeGeofenceErrorCode code;